+++ ecosystem/typescript/sdk_v2/src/type_utils/converter/genericConverter.ts
